# src/main/java/ru/brombin/incident_service/entity/ResponsibleService.java
package ru.brombin.incident_service.entity;

public enum ResponsibleService {
    HARDWARE_SERVICE,
    ADMIN_SERVICE,
    PRINT_SERVICE,
    DEVOPS_SERVICE,
    DEVELOPMENT_SERVICE,
    HOUSEKEEPING_SERVICE,
    PROCUREMENT_SERVICE,
    IT_SUPPORT,
    WAREHOUSE_SERVICE
}