# src/main/java/ru/brombin/incident_service/util/ErrorResponse.java
package ru.brombin.incident_service.util;

import lombok.AccessLevel;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.Setter;
import lombok.experimental.FieldDefaults;

@Setter
@Getter
@AllArgsConstructor
@FieldDefaults(level= AccessLevel.PRIVATE)
public class ErrorResponse {
    String message;
    Long timestamp;
}